Moved the PDF/OCR Sample to GA
